@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any
 from uuid import uuid4
 
 from aio_databases.record import Record
@@ -15,7 +15,7 @@ if TYPE_CHECKING:
 
 
 class Transaction(ABCTransaction):
-    savepoint: Optional[str] = None
+    savepoint: str | None = None
 
     async def _start(self) -> Any:
         connection = self.connection
@@ -46,35 +46,40 @@ class Connection(ABCConnection[TVConnection]):
     transaction_cls = Transaction
 
     async def _execute(self, query: str, *params, **options) -> tuple[int, Any]:
-        assert self._conn is not None
-        async with self._conn.cursor() as cursor:
+        conn = self._conn
+        assert conn is not None
+        async with conn.cursor() as cursor:  # type: ignore[missing-attribute]
             await cursor.execute(query, params, **options)
             return cursor.rowcount, cursor.lastrowid
 
     async def _executemany(self, query: str, *params, **options) -> Any:
-        assert self._conn is not None
-        async with self._conn.cursor() as cursor:
+        conn = self._conn
+        assert conn is not None
+        async with conn.cursor() as cursor:  # type: ignore[missing-attribute]
             await cursor.executemany(query, params, **options)
 
     async def _fetchall(self, query: str, *params, **options) -> list[TRecord]:
-        assert self._conn is not None
-        async with self._conn.cursor() as cursor:
+        conn = self._conn
+        assert conn is not None
+        async with conn.cursor() as cursor:  # type: ignore[missing-attribute]
             await cursor.execute(query, params, **options)
             rows = await cursor.fetchall()
             desc = cursor.description
             return [Record(row, desc) for row in rows]
 
     async def _fetchmany(self, size: int, query: str, *params, **options) -> list[TRecord]:
-        assert self._conn is not None
-        async with self._conn.cursor() as cursor:
+        conn = self._conn
+        assert conn is not None
+        async with conn.cursor() as cursor:  # type: ignore[missing-attribute]
             await cursor.execute(query, params, **options)
             rows = await cursor.fetchmany(size)
             desc = cursor.description
             return [Record(row, desc) for row in rows]
 
-    async def _fetchone(self, query: str, *params, **options) -> Optional[TRecord]:
-        assert self._conn is not None
-        async with self._conn.cursor() as cursor:
+    async def _fetchone(self, query: str, *params, **options) -> TRecord | None:
+        conn = self._conn
+        assert conn is not None
+        async with conn.cursor() as cursor:  # type: ignore[missing-attribute]
             await cursor.execute(query, params, **options)
             row = await cursor.fetchone()
             if row is None:
@@ -82,8 +87,9 @@ class Connection(ABCConnection[TVConnection]):
             return Record(row, cursor.description)
 
     async def _fetchval(self, query: str, *params, column: Any = 0, **options) -> Any:
-        assert self._conn is not None
-        async with self._conn.cursor() as cursor:
+        conn = self._conn
+        assert conn is not None
+        async with conn.cursor() as cursor:  # type: ignore[missing-attribute]
             await cursor.execute(query, params, **options)
             row = await cursor.fetchone()
             if row is None:
@@ -91,8 +97,9 @@ class Connection(ABCConnection[TVConnection]):
             return row[column]
 
     async def _iterate(self, query: str, *params, **options) -> AsyncIterator[Record]:
-        assert self._conn is not None
-        async with self._conn.cursor() as cursor:
+        conn = self._conn
+        assert conn is not None
+        async with conn.cursor() as cursor:  # type: ignore[missing-attribute]
             await cursor.execute(query, params, **options)
             desc = cursor.description
             while True:
@@ -113,9 +120,9 @@ class PGReplacer:
         return f"{match.group(1)}${self.num}"
 
 
-def pg_parse_status(status: str) -> Union[str, tuple[int, Any]]:
+def pg_parse_status(status: str) -> str | tuple[int, Any]:
     operation, params = status.split(" ", 1)
-    if operation in {"INSERT"}:
+    if operation == "INSERT":
         oid, rows = params.split()
         return int(rows), oid
 
